IN operator converted to = ANY operator
Fixed the IN operator to use the = ANY operator and
fixed a bug where an array with a NULL in it, ie [1, NULL, 3]
would ignore the NULL and output true or false. Comparison with
a NULL cascade to a NULL comparison for all chained comparisons.
This is true is SQL and Cypher.