Refactor Agtype builtin functions (#128)
Removed the 'age_' from the start of function names, this will
make error messages more clear. Also conforms to Postgres conventions
for linking sql function names to the c function that implements that
function.

the c code functions had the 'age_' replace with 'agtype_' because these
functions are for the agtype type and this follows the jsonb implementation
that we are building on top of.

Fixed a bug where startnode and endnode had to be input in cameCase only
to work in a cypher function.